requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
dnspython==1.16.0
eventlet==0.29.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
greenlet==0.4.17
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
sentry-sdk==0.19.1
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.11
Werkzeug==1.0.1
